+++ packages/terra-form-radio/CHANGELOG.md
@@ -2,6 +2,9 @@
 
 ## Unreleased
 
+* Fixed
+  * Fixed issue of `onChange` not triggered on first and last item of radio field while keyboard navigation.
+
 ## 4.39.0 - (August 11, 2023)
 
 * Changed